requirements.txt
streamlit==1.24.0
pandas==2.0.3
gspread==5.8.0
oauth2client==4.1.3
google-generativeai==0.1.0
tweepy==4.15.0
python-dotenv==1.0.0
